feat: Add corner keyword to pairplot method
Adds a `corner` boolean keyword to the `pairplot` method in the `amrex` module. When `corner` is `True`, only the lower triangle of the pairplot is plotted.

Includes a new test to verify the functionality of the `corner` keyword and refactors existing tests to use a shared data fixture.
